Changed querry to use long instead of int

+++ backEndDataStorage/SpringWebServer/src/main/java/com/edg/Dao/ServerDao.java
@@ -5,7 +5,7 @@ public interface ServerDao {
     void saveUserInDB(String sqlQuery, String userName, String passwordHash);
 
     String getAllTasks();
-    void saveTaskInDB(String sqlQuery, String taskName, String taskBody, boolean isCompleted, int timestamp);
+    void saveTaskInDB(String sqlQuery, String taskName, String taskBody, boolean isCompleted, long timestamp);
     void updateTaskInDB(String sqlQuery, String taskName, String taskBody, boolean isCompleted);
     void deleteTaskInDB(String sqlQuery, int taskID);
